Agregasi/requirements.txt
# Core
fastapi==0.110.3
uvicorn==0.34.0
gunicorn==23.0.0

# Data & Caching
redis==5.0.8
apify-client==2.3.0
pandas==2.2.2
pytz==2024.2

# Utilities
tenacity==9.0.0
pydantic-settings==2.8.1
python-dotenv==1.0.1
httpx==0.27.2

# Testing Dependencies
pytest==8.3.4
pytest-asyncio==0.21.2
pytest-cov==4.1.0
pytest-mock==3.12.0
fakeredis==2.20.0
